Broke up the tests into cleaner modules
